Fix the ImplicitStringLowerBound canonicalizer rule
