--- src/canonicalizer/rules/pending/implicit_string_lower_bound.h
@@ -1,25 +1,0 @@
-namespace sourcemeta::jsonbinpack::canonicalizer {
-
-class ImplicitStringLowerBound final : public sourcemeta::alterschema::Rule {
-public:
-  ImplicitStringLowerBound() : Rule("implicit_string_lower_bound"){};
-  [[nodiscard]] auto
-  condition(const sourcemeta::jsontoolkit::Value &schema,
-            const std::string &dialect,
-            const std::unordered_map<std::string, bool> &vocabularies) const
-      -> bool override {
-    return sourcemeta::jsontoolkit::schema::has_vocabulary<std::string>(
-               schema, vocabularies::validation) &&
-           schema.is_object() && schema.defines(keywords::validation::type) &&
-           schema.at(keywords::validation::type) == "string" &&
-           !schema.defines(keywords::validation::minLength);
-  }
-
-  auto transform(sourcemeta::jsontoolkit::JSON &document,
-                 sourcemeta::jsontoolkit::Value &value) const -> void override {
-    schema.assign(keywords::validation::minLength,
-                  static_cast<std::int64_t>(0));
-  }
-};
-
-} // namespace sourcemeta::jsonbinpack::canonicalizer
